wip move data up
